requirements.txt
flask==3.0.0
flask-socketio==5.3.6
flask-sqlalchemy==3.1.1
flask-cors==4.0.0
flask-migrate==4.0.5
python-dotenv==1.0.0
pyjwt==2.8.0
psycopg2-binary==2.9.9
langchain==1.2.8
langchain-openai==0.2.8
langchain-core==0.3.15
eventlet==0.36.1
gunicorn==21.2.0
